+++ Bootstrap/BootstrapTests/ACustomFolder/BootstrapCustomFolderTests.swift
@@ -1,16 +1,16 @@
-import Quick
 import Nimble
 import Nimble_Snapshots
+import Quick
 
 class BootstrapCustomFormatTests: QuickSpec {
     override func spec() {
-        describe("in some context", { () -> () in
+        describe("in some context") {
             var view: UIView!
 
             beforeEach {
                 setNimbleTestFolder("CustomFolder")
-                view = UIView(frame: CGRect(origin: CGPointZero, size: CGSize(width: 44, height: 44)))
-                view.backgroundColor = UIColor.blueColor()
+                view = UIView(frame: CGRect(origin: .zero, size: CGSize(width: 44, height: 44)))
+                view.backgroundColor = .blue
             }
 
             it("fails to find the snapshots due to the custom folder") {
@@ -20,10 +20,10 @@ class BootstrapCustomFormatTests: QuickSpec {
             it("finds the snapshots using a custom images directory") {
                 expect(view).to(haveValidSnapshot())
             }
-          
+
             it("finds device agnostic snapshots with custom images directory") {
                 expect(view).to(haveValidDeviceAgnosticSnapshot())
             }
-        })
+        }
     }
 }