+++ Bootstrap/BootstrapTests/DynamicTypeTests.swift
@@ -1,11 +1,11 @@
-import Quick
+import Bootstrap
 import Nimble
 import Nimble_Snapshots
-import Bootstrap
+import Quick
 
 class DynamicTypeTests: QuickSpec {
     override func spec() {
-        describe("in some context", {
+        describe("in some context") {
             var view: UIView!
 
             beforeEach {
@@ -15,8 +15,8 @@ class DynamicTypeTests: QuickSpec {
             }
 
             it("has a valid snapshot for all content size categories (iOS 9)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 9 && version.minorVersion == 3 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS9 else {
                     return
                 }
 
@@ -26,19 +26,19 @@ class DynamicTypeTests: QuickSpec {
             }
 
             it("has a valid snapshot for a single content size category (iOS 9)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 9 && version.minorVersion == 3 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS9 else {
                     return
                 }
 
-                expect(view).to(haveValidDynamicTypeSnapshot(sizes: [UIContentSizeCategoryExtraLarge]))
+                expect(view).to(haveValidDynamicTypeSnapshot(sizes: [.extraLarge]))
                 let name = "something custom_\(version.majorVersion)_\(version.minorVersion)"
-                expect(view).to(haveValidDynamicTypeSnapshot(named: name, sizes: [UIContentSizeCategoryExtraLarge]))
+                expect(view).to(haveValidDynamicTypeSnapshot(named: name, sizes: [.extraLarge]))
             }
 
             it("has a valid snapshot for all content size categories (iOS 10)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 10 && version.minorVersion == 0 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS10 else {
                     return
                 }
 
@@ -49,20 +49,20 @@ class DynamicTypeTests: QuickSpec {
             }
 
             it("has a valid snapshot for a single content size category (iOS 10)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 10 && version.minorVersion == 0 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS10 else {
                     return
                 }
 
-                expect(view).to(haveValidDynamicTypeSnapshot(sizes: [UIContentSizeCategoryExtraLarge]))
+                expect(view).to(haveValidDynamicTypeSnapshot(sizes: [.extraLarge]))
 
                 let name = "something custom_\(version.majorVersion)_\(version.minorVersion)"
-                expect(view).to(haveValidDynamicTypeSnapshot(named: name, sizes: [UIContentSizeCategoryExtraLarge]))
+                expect(view).to(haveValidDynamicTypeSnapshot(named: name, sizes: [.extraLarge]))
             }
 
             it("has a valid pretty-syntax snapshot (iOS 9)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 9 && version.minorVersion == 3 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS9 else {
                     return
                 }
 
@@ -70,8 +70,8 @@ class DynamicTypeTests: QuickSpec {
             }
 
             it("has a valid pretty-syntax snapshot (iOS 10)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 10 && version.minorVersion == 0 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS10 else {
                     return
                 }
 
@@ -79,27 +79,37 @@ class DynamicTypeTests: QuickSpec {
             }
 
             it("has a valid pretty-syntax snapshot for only one size category (iOS 9)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 9 && version.minorVersion == 3 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS9 else {
                     return
                 }
 
-                expect(view) == dynamicTypeSnapshot(sizes: [UIContentSizeCategoryExtraLarge])
+                expect(view) == dynamicTypeSnapshot(sizes: [.extraLarge])
             }
 
             it("has a valid pretty-syntax snapshot for only one size category (iOS 10)") {
-                let version = NSProcessInfo.processInfo().operatingSystemVersion
-                guard version.majorVersion == 10 && version.minorVersion == 0 else {
+                let version = ProcessInfo.processInfo.operatingSystemVersion
+                guard version.iOS10 else {
                     return
                 }
 
-                expect(view) == dynamicTypeSnapshot(sizes: [UIContentSizeCategoryExtraLarge])
+                expect(view) == dynamicTypeSnapshot(sizes: [.extraLarge])
             }
 
             it("has a valid snapshot with model and OS in name") {
                 expect(view).to(haveValidDynamicTypeSnapshot(isDeviceAgnostic: true))
                 expect(view).to(haveValidDynamicTypeSnapshot(named: "something custom with model and OS", isDeviceAgnostic: true))
             }
-        })
+        }
+    }
+}
+
+private extension OperatingSystemVersion {
+    var iOS10: Bool {
+        return majorVersion == 10 && minorVersion == 3
+    }
+
+    var iOS9: Bool {
+        return majorVersion == 9 && minorVersion == 3
     }
 }