+++ DynamicType/HaveValidDynamicTypeSnapshot.swift
@@ -1,44 +1,33 @@
-import Quick
-import Nimble
-import UIKit
 import FBSnapshotTestCase
+import Nimble
+import Quick
+import UIKit
 
 func allContentSizeCategories() -> [UIContentSizeCategory] {
-    #if swift(>=3.0)
-        return [
-            .extraSmall, .small, .medium,
-            .large, .extraLarge, .extraExtraLarge,
-            .extraExtraExtraLarge, .accessibilityMedium,
-            .accessibilityLarge, .accessibilityExtraLarge,
-            .accessibilityExtraExtraLarge, .accessibilityExtraExtraExtraLarge
-        ]
-    #else
-        return [
-            UIContentSizeCategoryExtraSmall, UIContentSizeCategorySmall, UIContentSizeCategoryMedium,
-            UIContentSizeCategoryLarge, UIContentSizeCategoryExtraLarge, UIContentSizeCategoryExtraExtraLarge,
-            UIContentSizeCategoryExtraExtraExtraLarge, UIContentSizeCategoryAccessibilityMedium,
-            UIContentSizeCategoryAccessibilityLarge, UIContentSizeCategoryAccessibilityExtraLarge,
-            UIContentSizeCategoryAccessibilityExtraExtraLarge, UIContentSizeCategoryAccessibilityExtraExtraExtraLarge
-        ]
-    #endif
+    return [
+        .extraSmall, .small, .medium,
+        .large, .extraLarge, .extraExtraLarge,
+        .extraExtraExtraLarge, .accessibilityMedium,
+        .accessibilityLarge, .accessibilityExtraLarge,
+        .accessibilityExtraExtraLarge, .accessibilityExtraExtraExtraLarge
+    ]
 }
 
 func shortCategoryName(_ category: UIContentSizeCategory) -> String {
-    #if swift(>=3.0)
-        return category.rawValue.replacingOccurrences(of: "UICTContentSizeCategory", with: "")
-    #else
-        return category.stringByReplacingOccurrencesOfString("UICTContentSizeCategory", withString: "")
-    #endif
+    return category.rawValue.replacingOccurrences(of: "UICTContentSizeCategory", with: "")
 }
 
-func combineMatchers<T>(_ matchers: [MatcherFunc<T>], ignoreFailures: Bool = false, deferred: (() -> Void)? = nil) -> MatcherFunc<T> {
+func combineMatchers<T>(_ matchers: [MatcherFunc<T>], ignoreFailures: Bool = false,
+                        deferred: (() -> Void)? = nil) -> MatcherFunc<T> {
     return MatcherFunc { actualExpression, failureMessage in
         defer {
             deferred?()
         }
 
         return try matchers.reduce(true) { acc, matcher -> Bool in
-            guard acc || ignoreFailures else { return false }
+            guard acc || ignoreFailures else {
+                return false
+            }
 
             let result = try matcher.matches(actualExpression, failureMessage: failureMessage)
             return result && acc
@@ -46,23 +35,23 @@ func combineMatchers<T>(_ matchers: [MatcherFunc<T>], ignoreFailures: Bool = fal
     }
 }
 
-public func haveValidDynamicTypeSnapshot(named name: String? = nil, usesDrawRect: Bool = false, tolerance: CGFloat? = nil, sizes: [UIContentSizeCategory] = allContentSizeCategories(), isDeviceAgnostic: Bool = false) -> MatcherFunc<Snapshotable> {
+public func haveValidDynamicTypeSnapshot(named name: String? = nil, usesDrawRect: Bool = false,
+                                         tolerance: CGFloat? = nil,
+                                         sizes: [UIContentSizeCategory] = allContentSizeCategories(),
+                                         isDeviceAgnostic: Bool = false) -> MatcherFunc<Snapshotable> {
     let mock = NBSMockedApplication()
 
     let matchers: [MatcherFunc<Snapshotable>] = sizes.map { category in
-        let sanitizedName = _sanitizedTestName(name)
+        let sanitizedName = sanitizedTestName(name)
         let nameWithCategory = "\(sanitizedName)_\(shortCategoryName(category))"
 
         return MatcherFunc { actualExpression, failureMessage in
-            #if swift(>=3.0)
-                mock.mockPrefferedContentSizeCategory(category)
-            #else
-                mock.mockPrefferedContentSizeCategory(category as String)
-            #endif
+            mock.mockPrefferedContentSizeCategory(category)
 
             let matcher: MatcherFunc<Snapshotable>
             if isDeviceAgnostic {
-                matcher = haveValidDeviceAgnosticSnapshot(named: nameWithCategory, usesDrawRect: usesDrawRect, tolerance: tolerance)
+                matcher = haveValidDeviceAgnosticSnapshot(named: nameWithCategory,
+                                                          usesDrawRect: usesDrawRect, tolerance: tolerance)
             } else {
                 matcher = haveValidSnapshot(named: nameWithCategory, usesDrawRect: usesDrawRect, tolerance: tolerance)
             }
@@ -76,19 +65,17 @@ public func haveValidDynamicTypeSnapshot(named name: String? = nil, usesDrawRect
     }
 }
 
-public func recordDynamicTypeSnapshot(named name: String? = nil, usesDrawRect: Bool = false, sizes: [UIContentSizeCategory] = allContentSizeCategories(), isDeviceAgnostic: Bool = false) -> MatcherFunc<Snapshotable> {
+public func recordDynamicTypeSnapshot(named name: String? = nil, usesDrawRect: Bool = false,
+                                      sizes: [UIContentSizeCategory] = allContentSizeCategories(),
+                                      isDeviceAgnostic: Bool = false) -> MatcherFunc<Snapshotable> {
     let mock = NBSMockedApplication()
 
     let matchers: [MatcherFunc<Snapshotable>] = sizes.map { category in
-        let sanitizedName = _sanitizedTestName(name)
+        let sanitizedName = sanitizedTestName(name)
         let nameWithCategory = "\(sanitizedName)_\(shortCategoryName(category))"
 
         return MatcherFunc { actualExpression, failureMessage in
-            #if swift(>=3.0)
-                mock.mockPrefferedContentSizeCategory(category)
-            #else
-                mock.mockPrefferedContentSizeCategory(category as String)
-            #endif
+            mock.mockPrefferedContentSizeCategory(category)
 
             let matcher: MatcherFunc<Snapshotable>
             if isDeviceAgnostic {